refs #24437 renamed multi plot files
